Move ID factories into the core package
